Improve sample. Move classes to a separate files
